folder structure modified to include webhook server
