+++ deploy/kubefledged-deployment.yaml
