--- pkg/webhook/server.go
@@ -1,220 +1,0 @@
-/*
-Copyright 2018 The kube-fledged authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package webhook
-
-import (
-	"crypto/tls"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"net/http"
-	"strings"
-
-	"github.com/golang/glog"
-	"k8s.io/api/admission/v1beta1"
-	admissionregistrationv1beta1 "k8s.io/api/admissionregistration/v1beta1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-)
-
-var (
-	runtimeScheme = runtime.NewScheme()
-	codecs        = serializer.NewCodecFactory(runtimeScheme)
-	deserializer  = codecs.UniversalDeserializer()
-	// (https://github.com/kubernetes/kubernetes/issues/57982)
-	defaulter = runtime.ObjectDefaulter(runtimeScheme)
-)
-
-var (
-	ignoredNamespaces = []string{
-		metav1.NamespaceSystem,
-		metav1.NamespacePublic,
-	}
-
-	requiredLabels = []string{
-		nameLabel,
-		instanceLabel,
-		versionLabel,
-		componentLabel,
-		partOfLabel,
-		managedByLabel,
-	}
-
-	addLabels = map[string]string{
-		nameLabel:      NA,
-		instanceLabel:  NA,
-		versionLabel:   NA,
-		componentLabel: NA,
-		partOfLabel:    NA,
-		managedByLabel: NA,
-	}
-)
-
-const (
-	admissionWebhookAnnotationValidateKey = "admission-webhook-example.banzaicloud.com/validate"
-	admissionWebhookAnnotationMutateKey   = "admission-webhook-example.banzaicloud.com/mutate"
-	admissionWebhookAnnotationStatusKey   = "admission-webhook-example.banzaicloud.com/status"
-	nameLabel                             = "app.kubernetes.io/name"
-	instanceLabel                         = "app.kubernetes.io/instance"
-	versionLabel                          = "app.kubernetes.io/version"
-	componentLabel                        = "app.kubernetes.io/component"
-	partOfLabel                           = "app.kubernetes.io/part-of"
-	managedByLabel                        = "app.kubernetes.io/managed-by"
-	// NA denotes Not applicable
-	NA = "not_available"
-)
-
-// Server denotes a webhook https server listening to validating and mutating webhooks from api-server
-type Server struct {
-	server *http.Server
-}
-
-type patchOperation struct {
-	Op    string      `json:"op"`
-	Path  string      `json:"path"`
-	Value interface{} `json:"value,omitempty"`
-}
-
-func init() {
-	_ = corev1.AddToScheme(runtimeScheme)
-	_ = admissionregistrationv1beta1.AddToScheme(runtimeScheme)
-	// defaulting with webhooks:
-	// https://github.com/kubernetes/kubernetes/issues/57982
-	// _ = v1.AddToScheme(runtimeScheme)
-}
-
-// CreateAndStartWebHookServer creates and starts and https webhook server
-func CreateAndStartWebHookServer(stopCh <-chan struct{}, port int) error {
-	certFile, keyFile := "", ""
-	pair, err := tls.LoadX509KeyPair(certFile, keyFile)
-	if err != nil {
-		glog.Errorf("Failed to load key pair: %v", err)
-		return err
-	}
-
-	whsvr := &Server{
-		server: &http.Server{
-			Addr:      fmt.Sprintf(":%v", port),
-			TLSConfig: &tls.Config{Certificates: []tls.Certificate{pair}},
-		},
-	}
-
-	// define http server and server handler
-	mux := http.NewServeMux()
-	mux.HandleFunc("/mutate", whsvr.serve)
-	mux.HandleFunc("/validate", whsvr.serve)
-	whsvr.server.Handler = mux
-
-	if err := whsvr.server.ListenAndServeTLS("", ""); err != nil {
-		glog.Errorf("Failed to listen and serve webhook server: %v", err)
-		return err
-	}
-
-	<-stopCh
-	glog.Info("Shutting down webhook server")
-	return nil
-}
-
-// Serve method for webhook server
-func (whsvr *Server) serve(w http.ResponseWriter, r *http.Request) {
-	var body []byte
-	if r.Body != nil {
-		if data, err := ioutil.ReadAll(r.Body); err == nil {
-			body = data
-		}
-	}
-
-	if len(body) == 0 {
-		glog.Error("empty body")
-		http.Error(w, "empty body", http.StatusBadRequest)
-		return
-	}
-
-	// verify the content type is accurate
-	contentType := r.Header.Get("Content-Type")
-	if contentType != "application/json" {
-		glog.Errorf("Content-Type=%s, expect application/json", contentType)
-		http.Error(w, "invalid Content-Type, expect `application/json`", http.StatusUnsupportedMediaType)
-		return
-	}
-
-	var admissionResponse *v1beta1.AdmissionResponse
-	ar := v1beta1.AdmissionReview{}
-	if _, _, err := deserializer.Decode(body, nil, &ar); err != nil {
-		glog.Errorf("Can't decode body: %v", err)
-		admissionResponse = &v1beta1.AdmissionResponse{
-			Result: &metav1.Status{
-				Message: err.Error(),
-			},
-		}
-	} else {
-		fmt.Println(r.URL.Path)
-		if r.URL.Path == "/mutate" {
-			admissionResponse = whsvr.mutate(&ar)
-		} else if r.URL.Path == "/validate" {
-			admissionResponse = whsvr.validate(&ar)
-		}
-	}
-
-	admissionReview := v1beta1.AdmissionReview{}
-	if admissionResponse != nil {
-		admissionReview.Response = admissionResponse
-		if ar.Request != nil {
-			admissionReview.Response.UID = ar.Request.UID
-		}
-	}
-
-	resp, err := json.Marshal(admissionReview)
-	if err != nil {
-		glog.Errorf("Can't encode response: %v", err)
-		http.Error(w, fmt.Sprintf("could not encode response: %v", err), http.StatusInternalServerError)
-	}
-
-	glog.Infof("Ready to write reponse ...")
-	if _, err := w.Write(resp); err != nil {
-		glog.Errorf("Can't write response: %v", err)
-		http.Error(w, fmt.Sprintf("could not write response: %v", err), http.StatusInternalServerError)
-	}
-
-}
-
-func admissionRequired(ignoredList []string, admissionAnnotationKey string, metadata *metav1.ObjectMeta) bool {
-	// skip special kubernetes system namespaces
-	for _, namespace := range ignoredList {
-		if metadata.Namespace == namespace {
-			glog.Infof("Skip validation for %v for it's in special namespace:%v", metadata.Name, metadata.Namespace)
-			return false
-		}
-	}
-
-	annotations := metadata.GetAnnotations()
-	if annotations == nil {
-		annotations = map[string]string{}
-	}
-
-	var required bool
-	switch strings.ToLower(annotations[admissionAnnotationKey]) {
-	default:
-		required = true
-	case "n", "no", "false", "off":
-		required = false
-	}
-
-	return required
-}
